backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
motor==3.3.2
aiofiles==23.2.1
email-validator==2.1.0
cors==1.0.1
fastapi-cors==0.0.6